+++ src/layout/Header/ToggleMenuList/index.jsx
@@ -1,0 +1,44 @@
+import { Link } from 'react-router-dom';
+import styled from 'styled-components';
+
+const ToggleMenuList = ({ toggleMenuRef, toggleMenuList, setIsToggleOpen }) => {
+  return (
+    <ToggleMenuListContainer ref={toggleMenuRef}>
+      {toggleMenuList.map(menu => {
+        return (
+          <Link key={menu.name} className='link-tag' to={menu.path} onClick={() => setIsToggleOpen(false)}>
+            {menu.name}
+          </Link>
+        );
+      })}
+    </ToggleMenuListContainer>
+  );
+};
+
+export default ToggleMenuList;
+
+const ToggleMenuListContainer = styled.div`
+  width: 12rem;
+  margin-top: 0.3rem;
+  margin-left: auto;
+  background-color: var(--toggle-background);
+  box-shadow: rgb(0 0 0 / 10%) 0px 0px 8px;
+
+  .link-tag {
+    display: block;
+    padding: 1.25rem;
+    font-size: 0.9rem;
+    cursor: pointer;
+
+    :nth-child(2) {
+      @media only screen and (min-width: 1024px) {
+        display: none;
+      }
+    }
+
+    :hover {
+      color: var(--a-tag-hover-text);
+      background-color: var(--a-tag-hover-background);
+    }
+  }
+`;
